Move construction controller into separate file
